Package Babelarr and add CLI
